pyproject.toml
[tool.poetry]
name = "fastapi-hashnode-api"
version = "0.1.0"
description = ""
authors = ["Frankie Rodriguez <rfrankie@gmail.com>"]
readme = "README.md"
packages = [{ include = "fastapi" }]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
redis = "^3.5.3"
gql = "^3.0.0"
httpx = "^0.19.0"
requests = "^2.28.2"
requests-toolbelt = "^0.10.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
